Bug fixes for connection closure
close_srvctcp() could wait indefinitely for outstanding acks to be
received and server_worked() to finish.  But we might never receive the
acks (e.g. they might be lost) and server_worker() could potentially
(but unlikely) be blocked.   So change to used timeouts for these
events, so that will definitely close the connection and terminate the
process in a finite amount of time.